Learning Path
This platform aims to let students build their own self-paced learning depending on their performance and understanding of the material. It lays the foundation and covers additional topics to ensure that students get the most out of their education. Not only that, children would be able to explore other areas of interests. Our target group is children aged 8 to 12 years old, with the goal of helping them choose the best route for them before starting middle school.